requirements.txt
streamlit>=1.31.0
anthropic>=0.8.0
google-generativeai>=0.3.0
openai>=1.12.0
python-dotenv>=1.0.0 